requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.0
numpy>=1.24.0
requests>=2.28.0
python-multipart>=0.0.6
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.2 